third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-delayed-discovery-service-found.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('DelayedServicesDiscoveryAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => gatt.getPrimaryService('heart_rate'))
    .then(services => {
      services = [].concat(services);
      for (let service of services)
        assert_equals(service.uuid, heart_rate.uuid);
    });
}, 'Request for service. Must return even when the services are not ' +
   'immediately discovered');

</script>
